moved cli.rs to main.rs
